Create README.md
Good Evening Nicholas! Here's what I got up to so far with the Java conversion.
I created a MyFrame Parent Class, and added a public JLabel to it.
I read xml from the top 40 and try to render it to the label with minimal success.

Your instruction, if you will accept this assignment... is to render this xml of latest Top40 results (preferably pretty printed), to the screen, either through the JLabel, or preferably a web view component.